Changes for greater compiler portability (in pll folder): 1. Calling fopen_s rather than fopen (in MSVC builds), in     pllAlignmentDataDumpFile (in alignment.c) 2.  Dynamically-sized arrays are (in MSVC builds) replaced     with arrays allocated via malloc, in vectorPopcount      (in fastDNAparsimony.c), broadcastAfterRateOpt,      reduceEvaluateIterative, execFunction (note, some of      execFunction's dynamic-sized on-stack arrays,      which aren't declared except in MPI builds, have not      been replaced with arrays allocated via malloc).      Likewise there's (MPI-only) code in distributeYVectors      that allocates dynamically-sized arrays on the stack,      that will still fail to compile, if you do an MPI build      under MSVC. 3.  Types of variable declarations switched to int where      necessary (to avoid warnings about implict casts, etc.) 4.  Returns of pointer differences explicitly cast to int.      For example, in saveSubtree() in topologies.c 5.  S_ISDIR and S_ISREG macros are now defined when      hardware.c is built under MSVC. 6. strcpy_s is used in place of strcpy in MSVC builds,      in pllHashAdd. 7. In MSVC builds, hash.c, newick.c, treeIO.c, utils.c all     define _CRT_SECURE_NO_WARNINGS (because they     do a lot of calls to sprintf that are hard to replace with     calls to sprintf_s). 8. In MSVC builds, struct timespec is *not* declared,     in pthread.h.

+++ pll/newviewGenericSpecial.c
@@ -2616,8 +2616,8 @@ void pllNewviewIterative (pllInstance *tr, partitionList *pr, int startIndex)
     for(model = 0; model < pr->numberOfPartitions; model++)
     {
       /* number of sites in this partition */
-      size_t            
-        width  = (size_t)pr->partitionData[model]->width;
+      int            
+        width  = pr->partitionData[model]->width;
 
       /* this conditional statement is exactly identical to what we do in pllEvaluateIterative */
 
@@ -2682,7 +2682,7 @@ void pllNewviewIterative (pllInstance *tr, partitionList *pr, int startIndex)
 
           /* get the number of states in the data stored in partition model */
           
-          states = (size_t)pr->partitionData[model]->states,
+          states = pr->partitionData[model]->states,
           
           /* get the length of the current likelihood array stored at node p. This is 
              important mainly for the SEV-based memory saving option described in here:
